+++ HISTORY.rst
@@ -1,4 +1,14 @@
-0.3.0 (2016-11-26)
+0.3.1 (2016-11-29)
+++++++++++++++++++
+
+- fixed bug: cmap of scatter
+- added taking into account scatter kwargs
+- fixed bug: xnptsmax cannot be smaller than xnpts
+- changed default value of xylim for graph, multigraph and scatter frames to (None, None, None, None)
+- fixed bug: xylim auto-determination when no data is on the graph
+- 
+
+0.3.0 (2016-11-28)
 ++++++++++++++++++
 
 - Added multi-lines graph-frames